Fix timezone and datetime naivety issue
datetime objects which instantiate without a tzinfo kwarg are naive, and thus cannot be compared with non-naive datetime objects. This was fixed by specifying that gmt_start_time and gmt_end_time in the Course object is in tzinfo=timezone.utc. As well, the time conversion did not account for Daylights Savings Time. This was fixed using python-dateutils, which was added as a dependency in requirements.txt