Hotfixes: Scheduled Events and /assignments command (#88)
* Merge with Paulmski/Discord-bot develop branch (#9)

* Decoupled fetch_due_dates and announce_assignments methods

* Removed python code quality checker.

* Fixed edge case

Fixed a bug where if the bot was started within the hour of 6am it would
not send out an announcement.

* Fixed bug

Fixed bug where labs or multiple class times would be counted as
seperate courses.

* Fixed bug where days_left were one day off

* Adjusted formatting for announce_assignments

* Remove time sensitive information after 1 hour

Co-authored-by: paulmski <paulmski@gmail.com>

* Update CONTRIBUTING.md

* Hotfix: Fixed typo with mentions (#87)

* Fixed typo with mentions

Co-authored-by: tuanyvan <93796037+tuanyvan@users.noreply.github.com>

* Fix timezone and datetime naivety issue

datetime objects which instantiate without a tzinfo kwarg are naive, and thus cannot be compared with non-naive datetime objects. This was fixed by specifying that gmt_start_time and gmt_end_time in the Course object is in tzinfo=timezone.utc. As well, the time conversion did not account for Daylights Savings Time. This was fixed using python-dateutils, which was added as a dependency in requirements.txt

* Add in ALL option for /assignments command

* Change embed color for announcement

Co-authored-by: paulmski <paulmski@gmail.com>