Hotfixes: Scheduled Events and /assignments command (#89)
* Fork merge (#7)

* Removed the use of @everyone

While it is important to notify users of events and assignments there
are members who are not taking all the same courses on the server.
Refraining from using @everyone gives more flexibility for people to
either ignore or look at announcements.

* Fixed poorly constructed class attribute  setters

* Review #69 Amendments (#70)

* Addressed review changes in #69 

* Removed absolute value function usage in `events.py` and its `days_left` if clauses.

* Enforcing single quote use where necessary

* Reinstated ABS

The ABS function is used to give a consistent appearance where late
assignments only have one "-" infront and all future assignments have a
single "+" infront.

* Created more flexible date parsing

* Removed !homework command

* Reinstating homework command

* Enforce naming convention for study group

* Enforce group naming convention

* Edited logic checking for !group

* Added Docstrings

* Added info logging

* Edited info logging

* Added version function

* Revised version of group command

1. Allows for naming a a study group without quotes and with multiple
words.
2. Sanitizes user input
3. Removes the need for the "-study-group" suffix

* Fixed strange formatting of homework announcement, and moved parsing
course_name and code to sheets_parser.py.

* Fixed formatting

* Change announcement range to -1 to 14 days.

* Removed extranious update.py

* Review #69 Amendments (#70)

* Addressed review changes in #69 

* Removed absolute value function usage in `events.py` and its `days_left` if clauses.

* Enforcing single quote use where necessary

* Reinstated ABS

The ABS function is used to give a consistent appearance where late
assignments only have one "-" infront and all future assignments have a
single "+" infront.

* Added Python continuous intigration code quality checker.

* Check for previous announcements in channel and delete them.

* Change version

* added !list courses functionality

* making sure i didnt change unnessecary stuff

* get rid of yucky print statements

* Modified !list course command and forced all codes to be upper case.

* Fixed docstring

Co-authored-by: Paul Molczanski <paulmski@gmail.com>
Co-authored-by: Michael Rosanelli <michaelrosanelli105@gmail.com>

* Changes in requirements.txt to py-cord

* Remove time sensitive information after 1 hour

* Merge from Paulmski/Discord-Bot main (#8)

* Decoupled fetch_due_dates and announce_assignments methods

* Removed python code quality checker.

* Fixed edge case

Fixed a bug where if the bot was started within the hour of 6am it would
not send out an announcement.

* Fixed bug

Fixed bug where labs or multiple class times would be counted as
seperate courses.

* Fixed bug where days_left were one day off

* Adjusted formatting for announce_assignments

Co-authored-by: paulmski <paulmski@gmail.com>

* Rollback changes in requirements.txt

* Changed announcements to show 7 days and modified description of events.

* Delete time sensitive information after 6 hours

* Change version

* Feature #72: Pycord migration to develop branch (#81)

* Feature #72: Migration to Pycord (#79)

* Merge with Paulmski/Discord-bot develop branch (#9)

* Decoupled fetch_due_dates and announce_assignments methods

* Removed python code quality checker.

* Fixed edge case

Fixed a bug where if the bot was started within the hour of 6am it would
not send out an announcement.

* Fixed bug

Fixed bug where labs or multiple class times would be counted as
seperate courses.

* Fixed bug where days_left were one day off

* Adjusted formatting for announce_assignments

* Remove time sensitive information after 1 hour

Co-authored-by: paulmski <paulmski@gmail.com>

* Merge fork develop into feature-72-pycord (#10)

* Decoupled fetch_due_dates and announce_assignments methods

* Removed python code quality checker.

* Fixed edge case

Fixed a bug where if the bot was started within the hour of 6am it would
not send out an announcement.

* Fixed bug

Fixed bug where labs or multiple class times would be counted as
seperate courses.

* Fixed bug where days_left were one day off

* Adjusted formatting for announce_assignments

* Remove time sensitive information after 1 hour

* Rollback changes in requirements.txt

* Changed announcements to show 7 days and modified description of events.

* Delete time sensitive information after 6 hours

* Change version

Co-authored-by: paulmski <paulmski@gmail.com>

* Changes to requirements.txt

* Started conversion to slash commands

* Fix coinflip command

* Fixed homework command

* Removal of homework command

* Fixed list command and renamed to assignments

This fix creates a COURSE_CODES constant which the  function uses to provide the valid choices for the command. The daily announcement function will have to be reworked because of the breaking changes to events.py to accomodate the functionality of the  function.

* Fixed assignments task loop

Assignment task loop fixed by allowing two object types (discord.channel.TextChannel and discord.ApplicationContext) into the announce_assignments() function

* Restore group command functionality

* Add timed response messages deletion

* Restore purge_study_groups functionality

* Changes to datetime formatting and announcement title

Replaced all instances of timedelta in events.py with datetime.now(timezone.utc), changed spacing in announcement embed title, and refactored string from time in Course object

* Renamed function/slash-command arguments

* Update README.md

* Update CONTRIBUTING.md

Co-authored-by: paulmski <paulmski@gmail.com>

* Feature #72: Pycord Migration - Group Command Logic (#80)

Co-authored-by: paulmski <paulmski@gmail.com>

* Completed #78

* Fixed bug where channels would never delete

* Minor changes to docs to use "/" instead of "!"

* Fixed #82

* Changed version to 1.1.0

* Grammatical changes to README.md

* updated requirements

* Revert dependencies to previous version (#72)

* Fixed typo with mentions

* Hotfixes: Scheduled Events and /assignments command (#88)

* Merge with Paulmski/Discord-bot develop branch (#9)

* Decoupled fetch_due_dates and announce_assignments methods

* Removed python code quality checker.

* Fixed edge case

Fixed a bug where if the bot was started within the hour of 6am it would
not send out an announcement.

* Fixed bug

Fixed bug where labs or multiple class times would be counted as
seperate courses.

* Fixed bug where days_left were one day off

* Adjusted formatting for announce_assignments

* Remove time sensitive information after 1 hour

Co-authored-by: paulmski <paulmski@gmail.com>

* Update CONTRIBUTING.md

* Hotfix: Fixed typo with mentions (#87)

* Fixed typo with mentions

Co-authored-by: tuanyvan <93796037+tuanyvan@users.noreply.github.com>

* Fix timezone and datetime naivety issue

datetime objects which instantiate without a tzinfo kwarg are naive, and thus cannot be compared with non-naive datetime objects. This was fixed by specifying that gmt_start_time and gmt_end_time in the Course object is in tzinfo=timezone.utc. As well, the time conversion did not account for Daylights Savings Time. This was fixed using python-dateutils, which was added as a dependency in requirements.txt

* Add in ALL option for /assignments command

* Change embed color for announcement

Co-authored-by: paulmski <paulmski@gmail.com>

Co-authored-by: tuanyvan <93796037+tuanyvan@users.noreply.github.com>
Co-authored-by: Michael Rosanelli <michaelrosanelli105@gmail.com>